requirements.txt
Django==3.1
psycopg2-binary==2.8.5
djangorestframework
Pillow
python-decouple
django-filter

